Move Array to its own module and add init doc
