+++ atom/scalars.py
@@ -6,7 +6,6 @@
 # The full license is in the file COPYING.txt, distributed with this software.
 #------------------------------------------------------------------------------
 from .catom import Member, DefaultValue, Validate, SetAttr
-from .coerced import Coerced
 
 
 class Value(Member):
@@ -197,23 +196,3 @@ class Unicode(Value):
             self.set_validate_mode(Validate.Unicode, None)
         else:
             self.set_validate_mode(Validate.UnicodePromote, None)
-
-
-class Array(Coerced):
-    """ A value of type `np.ndarray`
-
-    Values are coerced to ndarrays using np.array.
-
-    """
-    __slots__ = ()
-
-    def __init__(self, default=None, factory=None, kwargs=None):
-        import numpy as np
-        if factory is None:
-            if default is not None:
-                kwargs = kwargs or {}
-                factory = lambda: np.array(default, **kwargs)
-            else:
-                factory = lambda: ()
-        super(Array, self).__init__(
-                np.ndarray, factory=factory, coercer=np.array)
